packages/camera-analytics/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "camera-analytics"
version = "0.1.0"
description = "ObservAI camera analytics pipeline for real-time cafe insights"
requires-python = ">=3.9,<3.15"
dependencies = [
  "numpy>=1.23",
  "opencv-python>=4.8.0",
  "ultralytics>=8.1.0",
  "pyyaml>=6.0",
  "shapely>=2.0",
  "lapx>=0.5.2",
  "aiohttp>=3.9.1",
  "python-socketio[asyncio]>=5.11.0",
  "yt-dlp>=2025.12.8",
]

[project.optional-dependencies]
demographics = [
  "insightface>=0.7.3",
  "onnxruntime-gpu>=1.19.0",
  "timm>=0.9.0",
  "yacs>=0.1.8",
]
# TensorRT: NVIDIA GPU acceleration (YOLO + InsightFace ~2-3x speedup)
# Requires NVIDIA GPU with CUDA support. First run compiles engines (~2-5 min).
tensorrt = [
  "tensorrt-cu12>=10.0.0",
]
dev = [
  "black>=23.3.0",
  "mypy>=1.6.0",
  "ruff>=0.1.0",
  "pytest>=7.4.0",
  "pytest-timeout>=2.2.0",
  "pytest-benchmark>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["camera_analytics*"]
